chore: Update before_install.sh script to remove app.log file if it exists

+++ scripts/before_install.sh
@@ -42,4 +42,5 @@ export FROM_EMAIL_PASSWORD="${FROM_EMAIL_PASSWORD}"
 echo "Environment variables set"
 
 # Remove old log file
-rm /opt/vokki_cloud/app.log
+[ -f /opt/vokki_cloud/app.log ] && rm /opt/vokki_cloud/app.log
+
